refactor: lessons text
-Refactor lesson component text generation into a hook called useLessonText. Add final text into lessons data for Beginner, Intermediate, and Advanced texts.
