+++ client/src/components/layout/lessonpg/LoneSoneLone.tsx
@@ -1,0 +1,163 @@
+function LoneSoneLone() {
+  return (
+    <div className="p-8 text-black">
+      <h1 className="mb-4 text-2xl font-bold">
+        Typing Basics: Letters "a" and "s"
+      </h1>
+      <p className="mb-4">
+        Typing is an essential skill that will help you in many aspects of life,
+        from writing documents to chatting online. We're going to take it step
+        by step.
+      </p>
+      <div>
+        
+      </div>
+
+      <h2 className="mb-2 text-xl font-semibold">
+        Step 1: Understanding the Keyboard
+      </h2>
+      <p className="mb-4">
+        The keyboard is a tool that allows us to input text and commands into a
+        computer. Each key represents a letter, number, or symbol. The most
+        common layout is the QWERTY layout, named after the first six letters on
+        the top row of letters.
+      </p>
+
+      <h3 className="mb-2 text-lg font-semibold">Key Areas of the Keyboard:</h3>
+      <ul className="mb-4 ml-6 list-inside list-disc">
+        <li>
+          <strong>Home Row:</strong> This is the middle row of the letter keys
+          where your fingers will rest. It includes the keys: A, S, D, F, G, H,
+          J, K, L.
+        </li>
+        <li>
+          <strong>Home Position:</strong> This is where your fingers should rest
+          when you're not typing.
+        </li>
+      </ul>
+
+      <h2 className="mb-2 text-xl font-semibold">
+        Step 2: Positioning Your Fingers
+      </h2>
+      <ol className="mb-4 ml-6 list-inside list-decimal">
+        <li>
+          <strong>Sit Up Straight:</strong> Good posture helps with typing
+          efficiently.
+        </li>
+        <li>
+          <strong>Home Row Position:</strong> Place your left hand fingers on
+          the keys A, S, D, and F. Place your right hand fingers on the keys J,
+          K, L, and ; (semicolon).
+          <ul className="ml-6 list-inside list-disc">
+            <li>
+              Your left pinky should be on <strong>A</strong>.
+            </li>
+            <li>
+              Your left ring finger should be on <strong>S</strong>.
+            </li>
+            <li>
+              Your left middle finger should be on <strong>D</strong>.
+            </li>
+            <li>
+              Your left index finger should be on <strong>F</strong>.
+            </li>
+          </ul>
+        </li>
+      </ol>
+      <p className="mb-4">
+        Notice the small bump on the F and J keys. These help you find the home
+        row without looking.
+      </p>
+
+      <h2 className="mb-2 text-xl font-semibold">
+        Step 3: Typing the Letters "a" and "s"
+      </h2>
+      <h3 className="mb-2 ml-4 text-lg font-semibold">Pressing "a":</h3>
+      <ol className="mb-4 ml-10 list-inside list-decimal">
+        <li>Keep your left pinky on the "A" key.</li>
+        <li>Gently press down on the "A" key using your left pinky.</li>
+        <li>Release the key after pressing it.</li>
+      </ol>
+
+      <h3 className="mb-2 ml-4 text-lg font-semibold">Pressing "s":</h3>
+      <ol className="mb-4 ml-10 list-inside list-decimal">
+        <li>Keep your left ring finger on the "S" key.</li>
+        <li>Gently press down on the "S" key using your left ring finger.</li>
+        <li>Release the key after pressing it.</li>
+      </ol>
+
+      <h2 className="mb-2 text-xl font-semibold">Step 4: Practicing</h2>
+      <p className="mb-4">Let's practice typing "a" and "s."</p>
+
+      <h3 className="mb-2 ml-4 text-lg font-semibold">
+        Exercise 1: Typing "a"
+      </h3>
+      <ol className="mb-4 ml-10 list-inside list-decimal">
+        <li>Place your fingers in the home position.</li>
+        <li>Press the "A" key with your left pinky.</li>
+        <li>Return your pinky to the home position.</li>
+      </ol>
+      <p className="mb-4">
+        Repeat this 10 times:
+        <br />a a a a a a a a a a
+      </p>
+
+      <h3 className="mb-2 ml-4 text-lg font-semibold">
+        Exercise 2: Typing "s"
+      </h3>
+      <ol className="mb-4 ml-10 list-inside list-decimal">
+        <li>Place your fingers in the home position.</li>
+        <li>Press the "S" key with your left ring finger.</li>
+        <li>Return your ring finger to the home position.</li>
+      </ol>
+      <p className="mb-4">
+        Repeat this 10 times:
+        <br />s s s s s s s s s s
+      </p>
+
+      <h3 className="mb-2 ml-4 text-lg font-semibold">
+        Exercise 3: Alternating "a" and "s"
+      </h3>
+      <ol className="mb-4 ml-10 list-inside list-decimal">
+        <li>Place your fingers in the home position.</li>
+        <li>Press the "A" key with your left pinky.</li>
+        <li>Press the "S" key with your left ring finger.</li>
+        <li>
+          Return each finger to the home position after pressing each key.
+        </li>
+      </ol>
+      <p className="mb-4">
+        Repeat this 10 times:
+        <br />a s a s a s a s a s
+      </p>
+
+      <h2 className="mb-2 text-xl font-semibold">Tips for Success</h2>
+      <ul className="mb-4 ml-6 list-inside list-disc">
+        <li>
+          <strong>Look at the Screen:</strong> Try not to look at your hands
+          while typing. This helps build muscle memory.
+        </li>
+        <li>
+          <strong>Practice Regularly:</strong> The more you practice, the better
+          you will get. Even a few minutes each day can help.
+        </li>
+        <li>
+          <strong>Stay Relaxed:</strong> Keep your hands and fingers relaxed to
+          avoid fatigue.
+        </li>
+      </ul>
+
+      <h2 className="mb-2 text-xl font-semibold">Conclusion</h2>
+      <p className="mb-4">
+        Great job! You've just learned the basics of typing the letters "a" and
+        "s." As you continue to practice, you'll get faster and more accurate.
+        Remember, typing is a skill that takes time to develop, so be patient
+        with yourself and keep practicing. Once you're comfortable with "a" and
+        "s," we can move on to other letters and more advanced typing
+        techniques.
+      </p>
+    </div>
+  );
+}
+
+export default LoneSoneLone;
