--- client/src/data/LessonData.ts
@@ -1,803 +1,0 @@
-export type LessonDataType = {
-  id: string;
-  icon: string;
-  title: string;
-  lessonData: {
-    sectionTitle: string;
-    sectionId: string;
-    sectionData: { id: string; levelTitle: string }[];
-  }[];
-}[];
-
-export default function LessonData() {
-  //Array of obects used to manage lessons page structure
-  const pageData: LessonDataType = [
-    {
-      id: "beginner-id",
-      icon: "face",
-      title: "Beginner",
-      lessonData: [
-        {
-          sectionTitle: "Home Row Left Hand",
-          sectionId: "home-row-left-id",
-          sectionData: [
-            {
-              id: "as",
-              levelTitle: "as",
-            },
-            {
-              id: "ad",
-              levelTitle: "ad",
-            },
-            {
-              id: "af",
-              levelTitle: "af",
-            },
-            {
-              id: "sd",
-              levelTitle: "sd",
-            },
-            {
-              id: "fd",
-              levelTitle: "fd",
-            },
-            {
-              id: "asd",
-              levelTitle: "asd",
-            },
-            {
-              id: "fds",
-              levelTitle: "fds",
-            },
-            {
-              id: "asdf",
-              levelTitle: "asdf",
-            },
-            {
-              id: "asdf-capital",
-              levelTitle: "ASDF",
-            },
-            {
-              id: "asdfasdf",
-              levelTitle: "ASDFasdf",
-            },
-          ],
-        },
-        {
-          sectionTitle: "Home Row Right Hand",
-          sectionId: "home-row-right-id",
-          sectionData: [
-            {
-              id: "jk",
-              levelTitle: "jk",
-            },
-            {
-              id: "jl",
-              levelTitle: "jl",
-            },
-            {
-              id: "j;",
-              levelTitle: "j;",
-            },
-            {
-              id: "kl",
-              levelTitle: "kl",
-            },
-            {
-              id: ";k",
-              levelTitle: ";k",
-            },
-            {
-              id: "jkl",
-              levelTitle: "jkl",
-            },
-            {
-              id: ";lk",
-              levelTitle: ";lk",
-            },
-            {
-              id: "jkl:",
-              levelTitle: "JKL:",
-            },
-            {
-              id: "jkl:jkl;",
-              levelTitle: "JKL:jkl;",
-            },
-          ],
-        },
-        {
-          sectionTitle: "Home Row",
-          sectionId: "home-row-id",
-          sectionData: [
-            {
-              id: "asjk",
-              levelTitle: "asjk",
-            },
-            {
-              id: "adjl",
-              levelTitle: "adjl",
-            },
-            {
-              id: "afj;",
-              levelTitle: "afj;",
-            },
-            {
-              id: "sdkl",
-              levelTitle: "sdkl",
-            },
-            {
-              id: "fs;k",
-              levelTitle: "fs;k",
-            },
-            {
-              id: "asdjkl",
-              levelTitle: "asdjkl",
-            },
-            {
-              id: "fds;lk",
-              levelTitle: "fds;lk",
-            },
-            {
-              id: "asdfjkl:",
-              levelTitle: "ASDFJKL:",
-            },
-            {
-              id: "asdfasdfjkl:jkl;",
-              levelTitle: "ASDFasdfJKL:jkl;",
-            },
-          ],
-        },
-        {
-          sectionTitle: "Top Row Left Hand",
-          sectionId: "top-row-left-id",
-          sectionData: [
-            {
-              id: "qw",
-              levelTitle: "qw",
-            },
-            {
-              id: "qe",
-              levelTitle: "qe",
-            },
-            {
-              id: "qr",
-              levelTitle: "qr",
-            },
-            {
-              id: "qt",
-              levelTitle: "qt",
-            },
-            {
-              id: "wer",
-              levelTitle: "wer",
-            },
-            {
-              id: "qwe",
-              levelTitle: "qwe",
-            },
-            {
-              id: "ert",
-              levelTitle: "ert",
-            },
-            {
-              id: "qwert",
-              levelTitle: "qwert",
-            },
-            {
-              id: "qwert-capital",
-              levelTitle: "QWERT",
-            },
-          ],
-        },
-        {
-          sectionTitle: "Top Row Right Hand",
-          sectionId: "top-row-right-id",
-          sectionData: [
-            {
-              id: "yu",
-              levelTitle: "yu",
-            },
-            {
-              id: "yi",
-              levelTitle: "yi",
-            },
-            {
-              id: "yo",
-              levelTitle: "yo",
-            },
-            {
-              id: "yp",
-              levelTitle: "yp",
-            },
-            {
-              id: "uio",
-              levelTitle: "uio",
-            },
-            {
-              id: "yui",
-              levelTitle: "yui",
-            },
-            {
-              id: "iop",
-              levelTitle: "iop",
-            },
-            {
-              id: "yuiop",
-              levelTitle: "yuiop",
-            },
-            {
-              id: "yuiop-capital",
-              levelTitle: "YUIOP",
-            },
-          ],
-        },
-        {
-          sectionTitle: "Top Row",
-          sectionId: "top-row-id",
-          sectionData: [
-            {
-              id: "qwyu",
-              levelTitle: "qwyu",
-            },
-            {
-              id: "qeyi",
-              levelTitle: "qeyi",
-            },
-            {
-              id: "qryo",
-              levelTitle: "qryo",
-            },
-            {
-              id: "qtyp",
-              levelTitle: "qtyp",
-            },
-            {
-              id: "weruio",
-              levelTitle: "weruio",
-            },
-            {
-              id: "qweyui",
-              levelTitle: "qweyui",
-            },
-            {
-              id: "ertiop",
-              levelTitle: "ertiop",
-            },
-            {
-              id: "qwertyuiop",
-              levelTitle: "qwertyuiop",
-            },
-            {
-              id: "qwertyuiop-capital",
-              levelTitle: "QWERTYUIOP",
-            },
-            {
-              id: "qwertqwertyuiopyuiop",
-              levelTitle: "QWERTqwertYUIOPyuiop",
-            },
-          ],
-        },
-      ],
-    },
-    {
-      id: "intermediate-id",
-      icon: "face",
-      title: "Intermediate",
-      lessonData: [
-        {
-          sectionTitle: "Bottom Row Left Hand",
-          sectionId: "bottom-row-left-id",
-          sectionData: [
-            {
-              id: "zx",
-              levelTitle: "zx",
-            },
-            {
-              id: "zc",
-              levelTitle: "zc",
-            },
-            {
-              id: "zv",
-              levelTitle: "zv",
-            },
-            {
-              id: "zb",
-              levelTitle: "zb",
-            },
-            {
-              id: "xcv",
-              levelTitle: "xcv",
-            },
-            {
-              id: "zxc",
-              levelTitle: "zxc",
-            },
-            {
-              id: "cvb",
-              levelTitle: "cvb",
-            },
-            {
-              id: "zxcvb",
-              levelTitle: "zxcvb",
-            },
-            {
-              id: "zxcvb-capital",
-              levelTitle: "ZXCVB",
-            },
-          ],
-        },
-        {
-          sectionTitle: "Bottom Row Right Hand",
-          sectionId: "bottom-row-right-id",
-          sectionData: [
-            {
-              id: "nm",
-              levelTitle: "nm",
-            },
-            {
-              id: "n,",
-              levelTitle: "n,",
-            },
-            {
-              id: "n.",
-              levelTitle: "n.",
-            },
-            {
-              id: "n/",
-              levelTitle: "n/",
-            },
-            {
-              id: "m,.",
-              levelTitle: "m,.",
-            },
-            {
-              id: "nm,",
-              levelTitle: "nm,",
-            },
-            {
-              id: ",./",
-              levelTitle: ",./",
-            },
-            {
-              id: "nm,./",
-              levelTitle: "nm,./",
-            },
-            {
-              id: "nm<>?",
-              levelTitle: "NM<>?",
-            },
-          ],
-        },
-        {
-          sectionTitle: "Bottom Row",
-          sectionId: "bottom-row-id",
-          sectionData: [
-            {
-              id: "zxnm",
-              levelTitle: "zxnm",
-            },
-            {
-              id: "zcn,",
-              levelTitle: "zcn,",
-            },
-            {
-              id: "zvn.",
-              levelTitle: "zvn.",
-            },
-            {
-              id: "zbn/",
-              levelTitle: "zbn/",
-            },
-            {
-              id: "xcvm,.",
-              levelTitle: "xcvm,.",
-            },
-            {
-              id: "zxcnm,",
-              levelTitle: "zxcnm,",
-            },
-            {
-              id: "cvb,./",
-              levelTitle: "cvb,./",
-            },
-            {
-              id: "zxcvbnm,./",
-              levelTitle: "zxcvbnm,./",
-            },
-            {
-              id: "zxcvbnm,./-capital",
-              levelTitle: "ZXCVBNM,./",
-            },
-            {
-              id: "zxcvbzxcvbnm,./nm,./",
-              levelTitle: "ZXCVBzxcvbNM,./nm,./",
-            },
-          ],
-        },
-        {
-          sectionTitle: "English Words",
-          sectionId: "all-three-rows-id",
-          sectionData: [
-            {
-              id: "lower-case-1",
-              levelTitle: "lower case left hand",
-            },
-            {
-              id: "lower-case-2",
-              levelTitle: "lower case right hand",
-            },
-            {
-              id: "lower-case-3",
-              levelTitle: "lower case both hands",
-            },
-            {
-              id: "upper-case-1",
-              levelTitle: "UPPER CASE LEFT HAND",
-            },
-            {
-              id: "upper-case-2",
-              levelTitle: "UPPER CASE RIGHT HAND",
-            },
-            {
-              id: "upper-case-3",
-              levelTitle: "UPPER CASE BOTH HANDS",
-            },
-            {
-              id: "camel-case-1",
-              levelTitle: "Title Case Left Hand",
-            },
-            {
-              id: "camel-case-2",
-              levelTitle: "Title Case: Right Hand",
-            },
-            {
-              id: "camel-case-3",
-              levelTitle: "Title Case: Both Hands",
-            },
-            {
-              id: "pascal-case-1",
-              levelTitle: "MiXed CasE LeFt HaNd",
-            },
-            {
-              id: "pascal-case-2",
-              levelTitle: "MiXed CasE rIgHt HaNd",
-            },
-            {
-              id: "pascal-case-3",
-              levelTitle: "MiXed CasE BoTh HanDs",
-            },
-          ],
-        },
-      ],
-    },
-    {
-      id: "advanced-id",
-      icon: "face",
-      title: "Advanced",
-      lessonData: [
-        {
-          sectionTitle: "Number Row",
-          sectionId: "number-row-id",
-          sectionData: [
-            {
-              id: "number-row-left",
-              levelTitle: "123456",
-            },
-            {
-              id: "number-row-right",
-              levelTitle: "7890-=",
-            },
-            {
-              id: "number-row-full",
-              levelTitle: "1234567890-=",
-            },
-            {
-              id: "number-row-left-1",
-              levelTitle: "!@#$%^",
-            },
-            {
-              id: "number-row-right-1",
-              levelTitle: "&*()_+",
-            },
-            {
-              id: "number-left-full-1",
-              levelTitle: "123456!@#$%^",
-            },
-            {
-              id: "number-right-full-1",
-              levelTitle: "7890-=&*()_+",
-            },
-            {
-              id: "number-full-1",
-              levelTitle: "Full Number Row",
-            },
-          ],
-        },
-        {
-          sectionTitle: "Brackets",
-          sectionId: "brackets-id",
-          sectionData: [
-            {
-              id: "brackets-1",
-              levelTitle: "asdfjkl;{}",
-            },
-            {
-              id: "brackets-2",
-              levelTitle: "asdfjkl;[]",
-            },
-            {
-              id: "brackets-3",
-              levelTitle: "asdfjkl;{}[]",
-            },
-            {
-              id: "brackets-4",
-              levelTitle: "asdfjkl;()",
-            },
-            {
-              id: "brackets-5",
-              levelTitle: "asdfjkl;{}[]()",
-            },
-          ],
-        },
-        {
-          sectionTitle: "Symbols",
-          sectionId: "symbols-id",
-          sectionData: [
-            {
-              id: "symbol-top-left-row-1",
-              levelTitle: "`~!@#$%^",
-            },
-            {
-              id: "symbol-top-right-row-1",
-              levelTitle: "&*()_+-=",
-            },
-            {
-              id: "symbol-top-row-1",
-              levelTitle: "~!@#$%^&*()_+-=",
-            },
-            {
-              id: "symbol-right-right-col-1",
-              levelTitle: "=+[{]}|;':,<.>/?\\\"",
-            },
-            {
-              id: "symbol-all-1",
-              levelTitle: "`@#$%^&*()_~=+[{]}|;':,<.>/?\\\"",
-            },
-          ],
-        },
-        {
-          sectionTitle: "Letters, Numbers, & Symbols",
-          sectionId: "letters-nums-symbols-id",
-          sectionData: [
-            {
-              id: "symbol-words-quotes",
-              levelTitle: "Words In Quotes",
-            },
-            {
-              id: "symobol-numbers-quotes",
-              levelTitle: "Numbers In Quotes",
-            },
-            {
-              id: "symobol-everything-quotes",
-              levelTitle: "Letters, Numbers, Symbols",
-            },
-          ],
-        },
-        {
-          sectionTitle: "Tricky Words",
-          sectionId: "tricky-words-id",
-          sectionData: [
-            {
-              id: "tricky-words",
-              levelTitle: "Words",
-            },
-            {
-              id: "mixed-case",
-              levelTitle: "MiXed CasE",
-            },
-            {
-              id: "tricky-words-symbols",
-              levelTitle: "Words & Symbols",
-            },
-            {
-              id: "tricky-mixed-words-symbols",
-              levelTitle: "MiXed CaSe & Symbols",
-            },
-            {
-              id: "mixed-case-3",
-              levelTitle: "MiXed CaSe & All",
-            },
-          ],
-        },
-      ],
-    },
-    {
-      id: "graduation-id",
-      icon: "graduationHat",
-      title: "Graduation",
-      lessonData: [
-        {
-          sectionTitle: "You Made It",
-          sectionId: "you-made-it-id",
-          sectionData: [
-            {
-              id: "congratulations",
-              levelTitle: "Congratulations!",
-            },
-          ],
-        },
-      ],
-    },
-    {
-      id: "quotes-id",
-      icon: "quote",
-      title: "Quotes",
-      lessonData: [
-        {
-          sectionTitle: "Quote Section",
-          sectionId: "quote-lesson-id",
-          sectionData: [
-            {
-              id: "quote-1",
-              levelTitle: "Quote 1",
-            },
-          ],
-        },
-      ],
-    },
-    {
-      id: "novels-id",
-      icon: "book",
-      title: "Novels",
-      lessonData: [
-        {
-          sectionTitle: "First Book Theme",
-          sectionId: "first-book-theme-id",
-          sectionData: [
-            {
-              id: "first-book",
-              levelTitle: "First Book",
-            },
-          ],
-        },
-      ],
-    },
-    {
-      id: "insects-id",
-      icon: "bee",
-      title: "Insects",
-      lessonData: [
-        {
-          sectionTitle: "Flying Insects",
-          sectionId: "fyling-insects-id",
-          sectionData: [
-            {
-              id: "bumble-bees",
-              levelTitle: "Bumble Bees",
-            },
-          ],
-        },
-      ],
-    },
-    {
-      id: "animals-id",
-      icon: "paw",
-      title: "Animals",
-      lessonData: [
-        {
-          sectionTitle: "Animal 1",
-          sectionId: "Animal-1-id",
-          sectionData: [
-            {
-              id: "first-animal",
-              levelTitle: "🦑 First Animal",
-            },
-          ],
-        },
-      ],
-    },
-    {
-      id: "biology-id",
-      icon: "microscope",
-      title: "Biology",
-      lessonData: [
-        {
-          sectionTitle: "First Biology Section",
-          sectionId: "first-bio-lesson-id",
-          sectionData: [
-            {
-              id: "bio-text-1",
-              levelTitle: "Bio Text 1",
-            },
-          ],
-        },
-      ],
-    },
-    {
-      id: "history-id",
-      icon: "paperQuill",
-      title: "History",
-      lessonData: [
-        {
-          sectionTitle: "Canadian History",
-          sectionId: "canadian-history-lesson-id",
-          sectionData: [
-            {
-              id: "history-text-1",
-              levelTitle: "History 1",
-            },
-          ],
-        },
-      ],
-    },
-    {
-      id: "geography-id",
-      icon: "mountain",
-      title: "Geography",
-      lessonData: [
-        {
-          sectionTitle: "volcanoes",
-          sectionId: "volcanoes-lesson-id",
-          sectionData: [
-            {
-              id: "yellowstone-text-1",
-              levelTitle: "Yellowstone",
-            },
-          ],
-        },
-      ],
-    },
-    {
-      id: "math-id",
-      icon: "math",
-      title: "Math",
-      lessonData: [
-        {
-          sectionTitle: "Math",
-          sectionId: "math-lesson-id",
-          sectionData: [
-            {
-              id: "math-text-1",
-              levelTitle: "Math 1",
-            },
-          ],
-        },
-      ],
-    },
-    {
-      id: "law-id",
-      icon: "law",
-      title: "Law",
-      lessonData: [
-        {
-          sectionTitle: "Criminal Law",
-          sectionId: "criminal-law-lesson-id",
-          sectionData: [
-            {
-              id: "criminal-law-text-1",
-              levelTitle: "Criminal Law 1",
-            },
-          ],
-        },
-      ],
-    },
-    {
-      id: "business-id",
-      icon: "suitcase",
-      title: "Business",
-      lessonData: [
-        {
-          sectionTitle: "Finance",
-          sectionId: "finance-lesson-id",
-          sectionData: [
-            {
-              id: "investing-1",
-              levelTitle: "investing",
-            },
-          ],
-        },
-      ],
-    },
-  ];
-  return pageData;
-}
